Refactor: 분리된 도메인 로직 통합 (#29)
