+++ src/main/java/com/livable/server/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.livable.server.invitation.repository;
+package com.livable.server.member.repository;
 
 import com.livable.server.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
