+++ src/test/java/com/livable/server/invitation/service/InvitationServiceTest.java
@@ -5,9 +5,9 @@ import com.livable.server.core.response.ApiResponse.Success;
 import com.livable.server.invitation.domain.InvitationErrorCode;
 import com.livable.server.invitation.dto.InvitationResponse.AvailablePlacesDTO;
 import com.livable.server.invitation.dto.InvitationResponse.CommonPlaceDTO;
-import com.livable.server.invitation.repository.MemberRepository;
+import com.livable.server.member.repository.MemberRepository;
 import com.livable.server.invitation.repository.OfficeRepository;
-import com.livable.server.invitation.repository.ReservationRepository;
+import com.livable.server.reservation.repository.ReservationRepository;
 import com.livable.server.invitation.service.data.InvitationBasicData;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
